requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
discord.py==2.5.2
docopt==0.6.2
dotenv==0.9.9
frozenlist==1.6.0
idna==3.10
multidict==6.4.4
propcache==0.3.1
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.4.0
yarg==0.1.10
yarl==1.20.0
